pyproject.toml
# https://docs.astral.sh/uv/reference/settings/
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/

# PROJECT

[project]
name = "arcdemo"
version = "0.1.0"
description = "Workflow for call logs"
authors = [{ name = "ahwolf" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.0",
    "mlflow>=2.19.0",
    "numpy (==2.2.3)",
    "pandas (==2.2.3)",
    "plotly>=6.0.0",
    "scikit-learn>=1.6.1",
    "aiohappyeyeballs (==2.4.6)",
    "aiohttp (==3.11.12)",
    "aiosignal (==1.3.2)",
    "appnope (==0.1.4)",
    "asttokens (==3.0.0)",
    "attrs (==25.1.0)",
    "box (==0.1.5)",
    "cfgv (==3.4.0)",
    "columnar (==1.3.1)",
    "comm (==0.2.2)",
    "debugpy (==1.8.12)",
    "decorator (==5.1.1)",
    "distlib (==0.3.9)",
    "docutils (==0.21.2)",
    "executing (==0.8.2)",
    "filelock (==3.17.0)",
    "frozenlist (==1.5.0)",
    "geohash (==1.0)",
    "identify (==2.6.7)",
    "idna (==3.10)",
    "invoke (==2.2.0)",
    "ipykernel (==6.29.5)",
    "ipython (==8.32.0)",
    "jedi (==0.19.2)",
    "jupyter-client (==8.6.3)",
    "jupyter-core (==5.7.2)",
    "matplotlib-inline (==0.1.7)",
    "multidict (==6.1.0)",
    "nest-asyncio (==1.6.0)",
    "nodeenv (==1.9.1)",
    "packaging (==24.2)",
    "parso (==0.8.4)",
    "pexpect (==4.9.0)",
    "platformdirs (==4.3.6)",
    "pre-commit (==4.1.0)",
    "prompt-toolkit (==3.0.50)",
    "propcache (==0.2.1)",
    "psutil (==7.0.0)",
    "ptyprocess (==0.7.0)",
    "pure-eval (==0.2.3)",
    "pygments (==2.19.1)",
    "python-box (==7.3.2)",
    "python-dateutil (==2.8.2)",
    "python-geohash (==0.8.5)",
    "pytz (==2025.1)",
    "pyyaml (==6.0.2)",
    "pyzmq (==26.2.1)",
    "six (==1.17.0)",
    "timeago (==1.0.14)",
    "toolz (==1.0.0)",
    "tornado (==6.4.2)",
    "traitlets (==5.14.3)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2025.1)",
    "uv (==0.6.0)",
    "virtualenv (==20.29.2)",
    "wcwidth (==0.2.13)",
    "yarl (==1.18.3)",
]
license = { file = "LICENSE.txt" }
keywords = ["mlops", "python", "package"]

# LINKS

[project.urls]
Homepage = "https://github.com/ahwolf/arcdemo"
Documentation = "https://ahwolf.github.io/arcdemo/"
Repository = "https://github.com/ahwolf/arcdemoe"
"Bug Tracker" = "https://github.com/ahwolf/arcdemo/issues"
Changelog = "https://github.com/ahwolf/arcdemo/blob/main/CHANGELOG.md"

# SCRIPTS

[project.scripts]
arcdemo = "arcdemo.scripts:main"

# DEPENDENCIES

[dependency-groups]
checks = [
    "bandit>=1.8.0",
    "coverage>=7.6.8",
    "mypy>=1.13.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.6.1",
    "pytest>=8.3.3",
    "ruff>=0.8.1",
]
commits = ["commitizen>=4.0.0", "pre-commit>=4.0.1"]
dev = ["invoke>=2.2.0"]
docs = ["pdoc>=15.0.0"]
notebooks = ["ipykernel>=6.29.5", "nbformat>=5.10.4"]

# TOOLS

[tool.uv]
default-groups = ["checks", "commits", "dev", "docs", "notebooks"]

[tool.bandit]
targets = ["src"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["__main__.py", "src/arcdemo/pipeline/stage_01_data_generation.py", "src/arcdemo/pipeline/stage_02_data_preparation.py"]

[tool.mypy]
pretty = true
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--verbosity=2"
pythonpath = ["src"]

[tool.ruff]
fix = true
indent-width = 4
line-length = 100
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D100", "D103"]

# SYSTEMS


[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.20241230"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
